+++ pkg/aws/v1alpha3/cluster/validate_cluster.go
@@ -2,7 +2,11 @@ package cluster
 
 import (
 	"context"
+	"fmt"
+	"time"
 
+	"github.com/blang/semver/v4"
+	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -73,6 +77,16 @@ func (v *Validator) ValidateCreate(request *admissionv1.AdmissionRequest) (bool,
 		return false, microerror.Mask(err)
 	}
 
+	err = v.ClusterAnnotationUpgradeTimeIsValid(cluster)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = v.ClusterAnnotationUpgradeReleaseIsValid(cluster)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
 	return true, nil
 }
 
@@ -97,6 +111,16 @@ func (v *Validator) ValidateUpdate(request *admissionv1.AdmissionRequest) (bool,
 		return true, nil
 	}
 
+	err = v.ClusterAnnotationUpgradeTimeIsValid(cluster)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = v.ClusterAnnotationUpgradeReleaseIsValid(cluster)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
 	if v.isAdmin(request.UserInfo) || v.isInRestrictedGroup(request.UserInfo) {
 		err = v.ClusterStatusValid(oldCluster, cluster)
 		if err != nil {
@@ -117,6 +141,75 @@ func (v *Validator) ValidateUpdate(request *admissionv1.AdmissionRequest) (bool,
 	}
 
 	return true, nil
+}
+
+func (v *Validator) ClusterAnnotationUpgradeTimeIsValid(cluster *capiv1alpha3.Cluster) error {
+	if updateTime, ok := cluster.GetAnnotations()[annotation.UpdateScheduleTargetTime]; ok {
+		if !UpgradeScheduleTimeIsValid(updateTime) {
+			return microerror.Maskf(notAllowedError,
+				fmt.Sprintf("Cluster annotation '%s' value '%s' is not valid. Value must be in RFC822 format and should be a date in the future not more than 6 months away.",
+					annotation.UpdateScheduleTargetTime,
+					updateTime),
+			)
+		}
+	}
+	return nil
+}
+
+func UpgradeScheduleTimeIsValid(updateTime string) bool {
+	// parse time
+	t, err := time.Parse(time.RFC822, updateTime)
+	if err != nil {
+		return false
+	}
+	//time already passed
+	if t.Before(time.Now()) {
+		return false
+	}
+	// time too far in the future
+	if time.Until(t) > 4380*time.Hour {
+		return false
+	}
+	return true
+}
+
+func (v *Validator) ClusterAnnotationUpgradeReleaseIsValid(cluster *capiv1alpha3.Cluster) error {
+	if targetRelease, ok := cluster.GetAnnotations()[annotation.UpdateScheduleTargetRelease]; ok {
+		err := v.UpgradeScheduleReleaseIsValid(targetRelease, key.Release(cluster))
+		if err != nil {
+			return microerror.Maskf(notAllowedError,
+				fmt.Sprintf("Cluster annotation '%s' value '%s' is not valid. Value must be an existing giant swarm release version above the current release version %s and must not have a v prefix. %v",
+					annotation.UpdateScheduleTargetTime,
+					targetRelease,
+					key.Release(cluster),
+					err),
+			)
+		}
+	}
+	return nil
+}
+
+func (v *Validator) UpgradeScheduleReleaseIsValid(targetRelease string, currentRelease string) error {
+	// parse target version
+	t, err := semver.New(targetRelease)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+	// check if the release exists
+	_, err = aws.FetchRelease(&aws.Handler{K8sClient: v.k8sClient, Logger: v.logger}, t)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+	// parse current version
+	c, err := semver.New(currentRelease)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+	//check if target is higher than the current release
+	if t.LE(*c) {
+		return microerror.Maskf(notAllowedError, "Upgrade target release version has to be above current release version.")
+	}
+	return nil
 }
 
 func (v *Validator) ClusterLabelKeysValid(oldCluster *capiv1alpha3.Cluster, newCluster *capiv1alpha3.Cluster) error {